Refactor to use abstract class better
