+++ unityfiles/Assets/Scripts/GUI/ButtonChangeLevelState.cs
@@ -8,21 +8,12 @@ using UnityEngine.UI;
 /// It requires to be on an game object with Button script applied
 /// </summary>
 [RequireComponent(typeof(Button))]
-public class ButtonChangeLevelState : ButtonTrigger
+public class ButtonChangeLevelState : ButtonChangeStateTrigger
 {
     [SerializeField]
     private LEVEL_STATE levelState;
 
-    private void Awake() {
-        this.triggerButton = GetComponent<Button>();
-        this.triggerButton.onClick.AddListener(ChangeState);
-    }
-
-    private void OnDestroy() {
-        this.triggerButton.onClick.RemoveListener(ChangeState);
-    }
-
-    private void ChangeState(){
+    protected override void ChangeState(){
         LevelManager.Instance.ChangeLevelState(this.levelState);
     }
 }
