+++ unityfiles/Assets/Scripts/GUI/ButtonChangePauseMenuState.cs
@@ -2,20 +2,11 @@ using System;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class ButtonChangePauseMenuState : ButtonTrigger {
+public class ButtonChangePauseMenuState : ButtonChangeStateTrigger {
     [SerializeField]
     private PAUSE_MENU_STATE pauseState;
-
-    private void Awake() {
-        this.triggerButton = GetComponent<Button>();
-        this.triggerButton.onClick.AddListener(ChangeState);
-    }
-
-    private void OnDestroy() {
-        this.triggerButton.onClick.RemoveListener(ChangeState);
-    }
-
-    private void ChangeState() {
+    
+    protected override void ChangeState() {
         PauseMenu.Instance.ChangePauseMenuState(this.pauseState);
     }
 }
