--- unityfiles/Assets/Scripts/GUI/ButtonTrigger.cs
@@ -1,10 +1,0 @@
-using System;
-using UnityEngine;
-using UnityEngine.UI;
-
-public abstract class ButtonTrigger : MonoBehaviour {
-    /// <summary>
-    /// The button that triggers the click event
-    /// </summary>
-    protected Button triggerButton;
-}
